Refactor: Move matchMaterialsToSlots to machine-helpers
Extract the matchMaterialsToSlots helper function from MachinesSection
component to machine-helpers.ts for better organization and reusability.

- Move 56-line helper function to game logic layer
- Add MaterialSlot interface export
- Add JSDoc documentation
- Clean up unused imports in MachinesSection
- Reduce component file size from 386 to 326 lines
