+++ src/components/current-cell-info/MachinesSection.tsx
@@ -1,10 +1,6 @@
 import React from "react";
 import { useCellMap } from "../../game/CellMap";
-import {
-  InputMaterialWithQuantity,
-  Machine,
-  ParameterValues,
-} from "../../game/Machine";
+import { Machine, ParameterValues } from "../../game/Machine";
 import { MaterialInstance } from "../../game/Materials";
 import {
   operateMachineAction,
@@ -12,12 +8,11 @@ import {
   takeInputsFromMachineAction,
   takeOutputsFromMachineAction,
 } from "../../game/game-actions/player-actions";
-import { machineCanOperate } from "../../game/machine-helpers";
 import {
-  createMockMaterial,
-  getMaterialName,
-  materialMeetsInput,
-} from "../../game/material-helpers";
+  machineCanOperate,
+  matchMaterialsToSlots,
+} from "../../game/machine-helpers";
+import { createMockMaterial, getMaterialName } from "../../game/material-helpers";
 import {
   executeOperation,
   generateOperationPreview,
@@ -27,65 +22,6 @@ import {
 import { groupBy } from "../../utils/arrayUtils";
 import { useApplyGameAction, useGameState } from "../useGameState";
 import { MaterialIcon } from "./MaterialIcon";
-
-// Helper to match actual materials to operation requirements
-function matchMaterialsToSlots(
-  actualMaterials: MaterialInstance[],
-  requirements: ReadonlyArray<InputMaterialWithQuantity>
-): Array<{
-  requirement: InputMaterialWithQuantity;
-  material: MaterialInstance;
-  isValid: boolean;
-  isPlaceholder: boolean;
-}> {
-  const slots = [];
-  const availableMaterials = [...actualMaterials];
-
-  for (const requirement of requirements) {
-    for (let i = 0; i < requirement.quantity; i++) {
-      // Find a matching material
-      const materialIndex = availableMaterials.findIndex((material) =>
-        materialMeetsInput(material, requirement)
-      );
-
-      if (materialIndex !== -1) {
-        // Found a matching material
-        const material = availableMaterials[materialIndex];
-        availableMaterials.splice(materialIndex, 1);
-        slots.push({
-          requirement,
-          material,
-          isValid: true,
-          isPlaceholder: false,
-        });
-      } else {
-        // No matching material found - try to find any material for this slot
-        const anyMaterialIndex = availableMaterials.findIndex(() => true);
-        if (anyMaterialIndex !== -1) {
-          // Found a material but it doesn't match requirements
-          const material = availableMaterials[anyMaterialIndex];
-          availableMaterials.splice(anyMaterialIndex, 1);
-          slots.push({
-            requirement,
-            material,
-            isValid: false,
-            isPlaceholder: false,
-          });
-        } else {
-          // No material at all - show mock material as placeholder
-          slots.push({
-            requirement,
-            material: createMockMaterial(requirement),
-            isValid: false,
-            isPlaceholder: true,
-          });
-        }
-      }
-    }
-  }
-
-  return slots;
-}
 
 export const MachinesSection: React.FC = () => {
   const gameState = useGameState();
